Use correct date values in widget.
